+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - show photos from firebase
 - login
 - recognize moderator
+- introducing configurable components
 
 ## 0.2.2 (2018-11-05)
 - fix map iOS